+++ grails-app/resources/api/v1/IngredientResource.groovy
@@ -1,7 +1,6 @@
 package api.v1
 
 import grails.converters.JSON
-import mixology.Drink
 import mixology.Ingredient
 import mixology.IngredientService
 import mixology.User
@@ -9,7 +8,6 @@ import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.springframework.beans.factory.annotation.Autowired
 
-import javax.print.attribute.standard.Media
 import javax.ws.rs.DELETE
 import javax.ws.rs.GET
 import javax.ws.rs.POST
@@ -17,7 +15,6 @@ import javax.ws.rs.PUT
 import javax.ws.rs.Path
 import javax.ws.rs.PathParam
 import javax.ws.rs.Produces
-import javax.ws.rs.core.MediaType
 import javax.ws.rs.core.Response
 
 @Path('/v1/ingredients')
@@ -31,8 +28,9 @@ class IngredientResource extends BaseResource {
     @Produces('application/json')
     @GET
     public Response getAllIngredients() {
+        User user = getAuthenticatedUser()
         try {
-            def list = ingredientService.findAll()
+            def list = (user) ? ingredientService.findAll(user) : ingredientService.findAll()
             Response.ok( list ).build()
         } catch (Exception e) {
             Response.serverError().build()
@@ -43,10 +41,19 @@ class IngredientResource extends BaseResource {
     @Path('/{ingredientId}')
     @GET
     public Response getAnIngredient(@PathParam('ingredientId') Long ingredientId) {
+        User user = getAuthenticatedUser()
         try {
             def ingredient = Ingredient.findById(ingredientId)
-            if (!ingredient) throw NotFoundException("Ingredient not found using $ingredientId")
-            else Response.ok(ingredient).build()
+            if (!ingredient) badRequest("Ingredient not found using $ingredientId")
+            else {
+                if (!user) Response.ok(ingredient).build()
+                else {
+                    if (!user.ingredients.contains(ingredient)) badRequest("The user you provided does not have the ingredient you are inquiring about.")
+                    else {
+                        Response.ok(ingredient).build()
+                    }
+                }
+            }
         } catch (Exception e) {
             logger.error("There was an error finding the ingredient. Reason: ${e.getMessage()}")
             Response.serverError().build()
@@ -57,20 +64,23 @@ class IngredientResource extends BaseResource {
     @POST
     public Response createAnIngredient(Map params) {
         User user = getAuthenticatedUser()
-        Ingredient newIngredient
-        try {
-            newIngredient = new Ingredient(params)
-            if (newIngredient.validate()) {
-                ingredientService.save(newIngredient, user, false)
-                def message = """
+        if (!user) badRequest("A user must be provided")
+        else {
+            Ingredient newIngredient
+            try {
+                newIngredient = new Ingredient(params)
+                if (newIngredient.validate()) {
+                    ingredientService.save(newIngredient, user, false)
+                    def message = """
                     User ${user.firstName} ${user.lastName} added a new ingredient.
                     Ingredient ${newIngredient as JSON}
                 """
-                Response.ok(message).build()
-            } else throw new Exception("Invalid ingredient")
-        } catch (Exception e) {
-            logger.error("There was an exception while creating a new ingredient: ${e.message}")
-            badRequest("There was an exception while creating a new ingredient: ${e.message}")
+                    Response.ok(message).build()
+                } else throw new Exception("Invalid ingredient")
+            } catch (Exception e) {
+                logger.error("There was an exception while creating a new ingredient: ${e.message}")
+                badRequest("There was an exception while creating a new ingredient: ${e.message}")
+            }
         }
     }
 
@@ -79,48 +89,52 @@ class IngredientResource extends BaseResource {
     @PUT
     public Response updateAnIngredient(@PathParam('ingredientId') Long ingredientId, Map params) {
         User user = getAuthenticatedUser()
-        Ingredient updateIngredient
-        def message = ''
-        try {
-            updateIngredient = Ingredient.findById(ingredientId)
-            params.keySet().each { String prop ->
-                // only update property if not in list
-                // if user is admin user, we can update any value we wish
-                if ( !ignoreSpecificProperties(prop) || user.isAdmin() ) {
-                    def value = params."$prop"
-                    println "params.$prop = $value"
-                    try {
-                        updateIngredient."$prop" = value
-                    } catch (MissingPropertyException mpe) {
-                        logger.info("${mpe.class.simpleName} for Ingredient, property:$prop")
-                        message += "$prop is not a property of the Ingredient object.\n"
+        if (!user) badRequest("A user must be provided")
+        else {
+            Ingredient updateIngredient
+            def message = ''
+            try {
+                updateIngredient = Ingredient.findById(ingredientId)
+                params.keySet().each { String prop ->
+                    // only update property if not in list
+                    // if user is admin user, we can update any value we wish
+                    if ( !ignoreSpecificProperties(prop) || user.isAdmin() ) {
+                        def value = params."$prop"
+                        println "params.$prop = $value"
+                        try {
+                            updateIngredient."$prop" = value
+                        } catch (MissingPropertyException mpe) {
+                            logger.info("${mpe.class.simpleName} for Ingredient, property:$prop")
+                            message += "$prop is not a property of the Ingredient object.\n"
+                        }
+                    } else {
+                        message += "Cannot update property: $prop\n"
                     }
-                } else {
-                    message += "Cannot update property: $prop\n"
                 }
-            }
-            if (!updateIngredient) {
-                badRequest("No ingredient was found using the id: $ingredientId")
-            } else if ( !user.ingredients*.id.contains(ingredientId) ) {
-                badRequest("User does not have an ingredient using the id: $ingredientId")
-            }
-            if (updateIngredient.validate() && message == '') {
-                ingredientService.save(updateIngredient, user, false)
-                message = """
+                if (!updateIngredient) {
+                    badRequest("No ingredient was found using the id: $ingredientId")
+                } else if ( !user.ingredients*.id.contains(ingredientId) ) {
+                    badRequest("User does not have an ingredient using the id: $ingredientId")
+                }
+                if (updateIngredient.validate() && message == '') {
+                    ingredientService.save(updateIngredient, user, false)
+                    message = """
                     User ${user.firstName} ${user.lastName} updated their ingredient.
                     Ingredient ${updateIngredient as JSON}
                 """
-                logger.info(message)
-                Response.ok(updateIngredient).build()
+                    logger.info(message)
+                    Response.ok(updateIngredient).build()
+                }
+                else if (message != '') {
+                    message += "Please fix all errors to update your ingredient"
+                    badRequest(message)
+                }
+                else throw new Exception("Invalid ingredient")
             }
-            else if (message != '') {
-                message += "Please fix all errors to update your ingredient"
-                badRequest(message)
+            catch (Exception e) {
+                logger.error("There was an exception while updating a ingredient: ${e.message}")
+                badRequest("There was an exception while updating a ingredient: ${e.message}")
             }
-            else throw new Exception("Invalid ingredient")
-        } catch (Exception e) {
-            logger.error("There was an exception while updating a ingredient: ${e.message}")
-            badRequest("There was an exception while updating a ingredient: ${e.message}")
         }
     }
 
@@ -129,29 +143,32 @@ class IngredientResource extends BaseResource {
     @DELETE
     public Response deleteAnIngredient(@PathParam('ingredientId') Long ingredientId, Map params) {
         User user = getAuthenticatedUser()
-        Ingredient deleteIngredient
-        try {
-            def confirm
+        if (!user) badRequest("A user must be provided")
+        else {
+            Ingredient deleteIngredient
             try {
-                confirm = params?.Confirm
-            } catch (Exception e1) {
-                logger.error("There was an error while retrieving params value for Confirm: ${e1.message}")
-                confirm = 'No'
-            }
-            if (confirm != 'Yes') {
-                badRequest("Pass in Confirm:Yes to delete")
-            } else {
-                deleteIngredient = Ingredient.findById(ingredientId)
-                if (!deleteIngredient) {
-                    badRequest("There was no ingredient found with id: $ingredientId")
-                } else {
-                    ingredientService.delete(deleteIngredient.id, user, true)
-                    Response.ok("User ${user.firstName} ${user.lastName} deleted their ingredient.").build()
+                def confirm
+                try {
+                    confirm = params?.Confirm
+                } catch (Exception e1) {
+                    logger.error("There was an error while retrieving params value for Confirm: ${e1.message}")
+                    confirm = 'No'
                 }
+                if (confirm != 'Yes') {
+                    badRequest("Pass in Confirm:Yes to delete")
+                } else {
+                    deleteIngredient = Ingredient.findById(ingredientId)
+                    if (!deleteIngredient) {
+                        badRequest("There was no ingredient found with id: $ingredientId")
+                    } else {
+                        ingredientService.delete(deleteIngredient.id, user, true)
+                        Response.ok("User ${user.firstName} ${user.lastName} deleted their ingredient.").build()
+                    }
+                }
+            } catch (Exception e2) {
+                logger.error("There was an exception while trying to delete an ingredient:: ${e2.message}")
+                badRequest("There was an exception while trying to delete an ingredient:: ${e2.message}")
             }
-        } catch (Exception e2) {
-            logger.error("There was an exception while trying to delete an ingredient:: ${e2.message}")
-            badRequest("There was an exception while trying to delete an ingredient:: ${e2.message}")
         }
     }
 }
